requirements.txt
google-generativeai>=0.3.0
pdf2image>=1.16.3
Pillow>=10.0.0
pytesseract>=0.3.10
opencv-python-headless>=4.8.0
numpy>=1.24.0
PyYAML>=6.0
tenacity>=8.2.0
tqdm>=4.65.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0